Moved write_index_pages function and its tests from vis_recon_all to ciftify/html.py and its test script
