+++ bin/cifti_vis_recon_all.py
@@ -95,7 +95,7 @@ def main():
         return
 
     logger.info("Writing index pages to {}".format(settings.qc_dir))
-    write_index_pages(settings, qc_config)
+    ciftify.html.write_index_pages(settings, qc_config)
 
 def write_single_qc_page(settings, qc_config):
     qc_subdir = os.path.join(settings.qc_dir, settings.subject)
@@ -147,23 +147,6 @@ def modify_template_contents(template_contents, settings):
     modified_text = template_contents.replace('HCP_DATA_PATH', settings.hcp_dir)
     modified_text = modified_text.replace('SUBJID', settings.subject)
     return modified_text
-
-def write_index_pages(settings, qc_config):
-    subjects = ciftify.utilities.get_subj(settings.qc_dir)
-
-    index_html = os.path.join(settings.qc_dir, 'index.html')
-    with open(index_html, 'w') as index_page:
-        ciftify.html.add_page_header(index_page, qc_config,
-                settings.qc_mode, active_link='index.html')
-        ciftify.html.add_image_and_subject_index(index_page, qc_config.images,
-                subjects, settings.qc_mode)
-
-    for image in qc_config.images:
-        if not image.make_index:
-            continue
-        title = "{} View Index ({} space)".format(image.name, settings.qc_mode)
-        ciftify.html.write_image_index(settings.qc_dir, subjects, qc_config,
-                settings.qc_mode, image.name, title=title)
 
 if __name__ == "__main__":
     main()
